requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform>=1.38.0
pillow
numpy
requests
